feat(Modal): Implement autofocus support (fixes #55)
When a modal is shown, its root element is given focus so keyboard events can be handled. This change allows developers to use the HTML5 autofocus attribute to indicate that an input element should be given focus, instead of the root element.

This also involved propagating the autofocus attribute through a couple of layers of for/input elements.