feat(Modal): Implement openAction (refs #55)
openAction will fire just after the modal has been shown, but before the transition animation, and openedAction will fire after the modal is completely shown, including transition animation.

The idea is that openAction can be used to access and manipulate the DOM at a time when it's no longer display: none, so size calculations and the like can be performed, and openedAction can be used to perform actions after the entire show animation is complete, such as setting focus, starting custom in-modal animations, etc.

Note that this technically breaks backwards compatibility, as it changes the timing of when openedAction fires.