Perform cleanup with modal willRemoveElement
The removal of the `modal-open` class is handled in the `hideModal`
function which also performs proper math to remove padding that was
previously added when the modal opened. Since this happens when `hide()`
is called (via the `open` property observer) everything works. But if
you programmatically remove the modal via ember *without* using the
`open` property but using `{{#if …}}` like the docs impart the cleanup
is not performed.

Unfortunately the `willRemoveElement` is the only hook and it does not
perform the same cleanup that `hide` does and the result is that the
scroll bar padding is not reset.

The result was that closing the modal via Ember from the outside of
the component caused the right side of the page to squish
(`padding-right`) and be compounded each time a modal closes from ember.
This was noticed in Firefox but Chrome was not affected (I have no idea
why).

An example to reproduce this:

    {{#if showModal}}
      {{#bs-modal title="Simple Dialog"}}
        Hello world!
      {{/bs-modal}}
    {{/if}}

    export default Ember.Component.extend({
      showModal: false,

      infinateModalLoop: Ember.on('didInsertElement', function () {
        this.toggleProperty('showModal');
        Ember.run.later(this, 'infinateModalLoop', 200);
      })
    });