Relates to #155
Relates to #155

As I've explained in #155, the modal dialog is not closing anymore. Maybe, if you have issues with glimmer2 and you are adding these `div`s inbecause of these, then there can be a way to detect the glimmer version? Modals in Ember `1.13.10` are definitely, when adding these `div`s.