fix(Build): fix build issue with ember-cli 2.12.0. Thanks to @rwjblue
This fixes issues with ember-cli@2.12, and works on prior versions
of ember-cli. Simply put, the concept that we are trying to do here
is to filter down the set of things in the tree before we process
them.

This will also have the positive result of not transpiling all of
the compoents and templates for both bootstrap versions, just before
removing one of them.