feat(jQuery): Removed jQuery-dependency
Refactored all components to use native DOM instead of jQuery. This let's you build jQuery-less Ember apps. See [this blog post](http://miguelcamba.com/blog/2017/04/11/the-future-of-embers-testing-and-the-beheading-of-jquery/) for more context.
Closes #300 

BREAKING CHANGE:
The `triggerElement` and `viewportSelector` properties of tooltips and popovers now expect a valid CSS Selector. If you used any jQuery specific non-standard selector features like `:eq()` for example, this needs to be refactored!