feat(Modal): Use the Glimmer `in-element` API
For Glimmer2 based apps (Ember >= 2.10) this will use the `in-element` API to render modals at the root of the DOM, instead of ember-wormhole. This will be more performant, and reduce app size by not bringing in the ember-wormhole dependency. Only for older Ember versions it will fall back to ember-wormhole.