feat(Button): optional prevent concurrent click actions
Adds a `preventConcurrency` option to `{{bs-button}}`. If `true`, clicking
a button will not fire `onClick` action if a promise returned by an earlier
execution is still pending.

This allows to replace ember-concurrency for simple use cases. It's
basically the same as using an ember-concurrency task with `drop()`
modifier as `onClick` action.

Part of #638