Hide all validations after form submissions
After a successful form submission (i.e. no validation errors) which changes some state (e.g. creates a record), this could make the model's validation invalid (think of an `exclude` validation that a property is unique when creating an model, which becomes invalid after the form has been submitted and the new name has been added to a live updating collection). Previously if the form remains visible this would show a new validation error, after a successful form submission, because the form elements `showOwnValidation` has been activated before!

With this change, the display of validations is resetted after a successful submission, as if a fresh form component was rendered, so the user would have to resubmit or trigger the element's validation display (focusout) again to show the validation error that has been added meanwhile.