FormElement: revert passing HTML attributes to its control component w/ angle brackets
As we need to also apply HTML attributes to the form elements `.form.group`, e.g. for test selectors, while not having those attributes applied twice, we don't pass them to the control component anymore. To apply HTML attributes to the control, they should be passed to the yielded component:

```hbs
<form.element @property="name" as |el|>
  <el.control placeholder="Enter your name" />
</form.element>
```

Fixes #844