Render Modals, Tooltips and Popovers in a wormhole even in tests
Previously those components would fallback to `renderInPlace=true` if the wormhole destination is not found, which was the case in tests. Now they will render wormholed to the test container element, so they work just as they would in the real app. This might break some tests, if they make assumptions about the place in DOM of those components in tests.