reflect bootstrap's behaviour on mouse down and mouse up, fixes #934
In bootstrap's source code, it tries to not close the modals if the
mouse down event was inside of the modal and mouse up event was outside
of the modal. With these changes we are copying the same behaviour from
it. We are also using similar variables to reflect it's source code.

Here is the click event of bootstrap: https://github.com/twbs/bootstrap/blob/master/js/src/modal.js#L361
Here is the handlers for mouse down and up events: https://github.com/twbs/bootstrap/blob/master/js/src/modal.js#L154