support optional feature default-async-observers
Adds support for optional feature `default-async-observers: true` by
explicitly setting all observers to sync, which must not be run async.

It also enables optional feature `template-only-glimmer-components` for
testing. It was already supported. Only changed it to prevent a regression.

Adds an additional ember try scenario that ensures it's working for the
opposite optional feature configuration. This is similar to the
ember-classic scenario, which is part of Ember CLI's blueprints since
3.15.

Closes #939
Superseds #945