support optional feature default-async-observers (#951)
Adds support for optional feature `default-async-observers: true` by explicitly setting all observers to sync, which must not be run async.

It also enables optional feature `template-only-glimmer-components` for testing. It was already supported. Only changed it to prevent a regression.

Adds an additional ember try scenario that ensures it's working for the opposite optional feature configuration. This is similar to the ember-classic scenario, which is part of Ember CLI's blueprints since 3.15.

Tests are failing if optional feature application-template-wrapper is disabled. It's commented out in ember try configuration therefor. This should be an issue only with our test setup. See https://github.com/kaliber5/ember-bootstrap/pull/951#discussion_r363103965 for details.

Closes #939
Superseds #945