Fix CI config for docs
The previous `if` condition did not work correctly, resulting in a skipped job. Note this will only run when a release in Github is published, not merely a git tag. But as we use `release-it` this will happen automatically. Added the `edited` event type just to be able to manually trigger the (previously skipped) job run. Hopefully this works now...