Add experimental/incomplete Bootstrap 5 support to existing components
This change *pretends* that BS5 components require the same markup as BS4. This should be true for many components, but not all. Follow up PRs will be created to address the later cases, see #1494.

But at least this is better than what we had before, as the macros checks for only 4 or 5 would leave the markup for BS5 broken basically everywhere. And allows us to iterate step by step on fixing the remaining issues, with BS5 already enjoying full CI coverage.