bs-dropdown dynamic positioning should not be used when in a navbar (#1946)
 bs-dropdown dynamic positioning should n
fixes https://github.com/ember-bootstrap/ember-bootstrap/issues/1945

Bootstrap 5 (and 4) do not use popper to position dropdowns that are in navs. The bootstrap dropdown js detects if it is in a nav or specified to be static and instructs Popper to not apply any styles ([here's the dropdown disable code](https://github.com/twbs/bootstrap/blob/8fcfce1ebd988fd4e5ed2f51acaa5c49e6b301c9/js/src/dropdown.js#L312-L319)).

This change uses the existing `inNav` flag and passes it to the dropdown menu component and uses that to disable popper's styles in the same method as bootstrap.