Convert docs components to native classes (#2087)
ESLint was warning about the usage of classic Ember syntax. 
The following files have been updated to the native class syntax:

* `controllers/demo.js`
* `controllers/demo/alert.js`
* `controllers/demo/carousel.js`
* `controllers/demo/collapse.js`
* `controllers/changelog.js`
* `models/login.js`
* `routes/demo.js`
* `routes/changelog.js`
* `services/component.js`

Changelog has been fixed to be compatible with `ember-cli-showdown` 8.x.

`controllers/application.js` has been removed since it's only use was checking for `/acceptance` in the url, which was impossible to happen.