Fix click events inside `bs-dropdown`s not activating
Due to a difference between classic Ember components and Glimmer components, the render timing of the `isOpen` variable of BsDropdown changed.
This caused the dropdown to close before a click handler could be activated, which in turn made it impossible to add custom actions to BsDropdown menu items.
By deferring the `this.isOpen = false` to the next runloop iteration, the click handler gets the time to fire before the BsDropdown disappears.