Tests: Add support for writing tests in Gjs/Gts
When changing the tests to get rid of the action helpers, it might be a good time to also update the tests to Gjs/Gts.
The changes for the action helper will be influencing how we access scope variables in the tests, and so will using ember-template-import syntax.
The commit prepares the test-app for template-imports and converts one file to gjs and one to gts as a proof-of-concept and to verify that the configuration of tools such as ESLint and Prettier still work correctly.